c: shadows

+++ c/camera.h
@@ -8,14 +8,14 @@
 typedef struct camera_t {
 	int hsize;
 	int vsize;
-	float fov;
+	double fov;
 	matrix4_t transform;
-	float pixel_size;
-	float half_width;
-	float half_height;
+	double pixel_size;
+	double half_width;
+	double half_height;
 } camera_t;
 
-camera_t camera(int hsize, int vsize, float tfov);
+camera_t camera(int hsize, int vsize, double tfov);
 ray_t camera_ray_for_pixel(camera_t c, int x, int y);
 canvas_t camera_render(camera_t c, world_t w);
