+++ c/ray.h
@@ -9,7 +9,7 @@ typedef struct ray_t {
 } ray_t;
 
 ray_t ray(vec4_t origin, vec4_t direction);
-vec4_t ray_position(ray_t r, float t);
+vec4_t ray_position(ray_t r, double t);
 ray_t ray_transform(ray_t r, matrix4_t m);
 
 void ray_demo();
