+++ c/transform.h
@@ -2,12 +2,12 @@
 
 #include "matrix4.h"
 
-matrix4_t translation(float x, float y, float z);
-matrix4_t scaling(float x, float y, float z);
-matrix4_t rotate_x(float r);
-matrix4_t rotate_y(float r);
-matrix4_t rotate_z(float r);
-matrix4_t shearing(float a, float b, float c, float d, float e, float f);
+matrix4_t translation(double x, double y, double z);
+matrix4_t scaling(double x, double y, double z);
+matrix4_t rotate_x(double r);
+matrix4_t rotate_y(double r);
+matrix4_t rotate_z(double r);
+matrix4_t shearing(double a, double b, double c, double d, double e, double f);
 
 matrix4_t transform(matrix4_t *ms, int n);
